requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.10.3
djangorestframework==3.5.3
enum34==1.1.6
funcsigs==1.0.2
ipython==5.1.0
ipython-genutils==0.1.0
mock==2.0.0
nltk==3.2.1
numpy==1.11.2
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.9
protobuf==3.0.0
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
